# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CDN Admin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AuthUserRole = (typeof AuthUserRole)[keyof typeof AuthUserRole];

export const AuthUserRole = {
  superadmin: "superadmin",
  admin: "admin",
  user: "user",
} as const;

export interface AuthUser {
  id: string;
  username?: string | null;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  profileImage?: string | null;
  profileImageUrl?: string | null;
  role: AuthUserRole;
}

export interface GetCurrentAuthUserResponse {
  authenticated: boolean;
  user?: AuthUser | null;
}

export interface ExchangeMobileAuthorizationCodeBody {
  code: string;
  code_verifier: string;
  redirect_uri: string;
  state: string;
  nonce?: string | null;
}

export interface ExchangeMobileAuthorizationCodeResponse {
  token: string;
}

export interface LogoutMobileSessionResponse {
  success: boolean;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  superadmin: "superadmin",
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: string;
  username: string;
  firstName?: string;
  lastName?: string;
  profileImage?: string | null;
  role: UserRole;
  createdAt: string;
}

export type UpdateUserBodyRole =
  (typeof UpdateUserBodyRole)[keyof typeof UpdateUserBodyRole];

export const UpdateUserBodyRole = {
  superadmin: "superadmin",
  admin: "admin",
  user: "user",
} as const;

export interface UpdateUserBody {
  role: UpdateUserBodyRole;
}

export type DomainStatus = (typeof DomainStatus)[keyof typeof DomainStatus];

export const DomainStatus = {
  active: "active",
  inactive: "inactive",
  provisioning: "provisioning",
} as const;

export interface Domain {
  id: number;
  name: string;
  sslEnabled: boolean;
  cacheEnabled: boolean;
  /** Cache TTL in seconds */
  cacheTtl: number;
  status: DomainStatus;
  createdAt: string;
  updatedAt: string;
}

export type OriginProtocol =
  (typeof OriginProtocol)[keyof typeof OriginProtocol];

export const OriginProtocol = {
  http: "http",
  https: "https",
} as const;

export interface Origin {
  id: number;
  domainId: number;
  /** IP address or hostname of the origin server */
  address: string;
  port: number;
  protocol: OriginProtocol;
  /** Load balancing weight (1-100) */
  weight: number;
  isActive: boolean;
  createdAt: string;
}

export type DomainWithOrigins = Domain & {
  origins: Origin[];
};

export interface CreateDomainBody {
  name: string;
  sslEnabled?: boolean;
  cacheEnabled?: boolean;
  cacheTtl?: number;
}

export type UpdateDomainBodyStatus =
  (typeof UpdateDomainBodyStatus)[keyof typeof UpdateDomainBodyStatus];

export const UpdateDomainBodyStatus = {
  active: "active",
  inactive: "inactive",
  provisioning: "provisioning",
} as const;

export interface UpdateDomainBody {
  name?: string;
  sslEnabled?: boolean;
  cacheEnabled?: boolean;
  cacheTtl?: number;
  status?: UpdateDomainBodyStatus;
}

export type CreateOriginBodyProtocol =
  (typeof CreateOriginBodyProtocol)[keyof typeof CreateOriginBodyProtocol];

export const CreateOriginBodyProtocol = {
  http: "http",
  https: "https",
} as const;

export interface CreateOriginBody {
  address: string;
  port?: number;
  protocol?: CreateOriginBodyProtocol;
  weight?: number;
}

export type UpdateOriginBodyProtocol =
  (typeof UpdateOriginBodyProtocol)[keyof typeof UpdateOriginBodyProtocol];

export const UpdateOriginBodyProtocol = {
  http: "http",
  https: "https",
} as const;

export interface UpdateOriginBody {
  address?: string;
  port?: number;
  protocol?: UpdateOriginBodyProtocol;
  weight?: number;
  isActive?: boolean;
}

export interface PurgeCacheBody {
  /** Specific path to purge, or omit to purge all */
  path?: string;
}

export interface PurgeCacheResponse {
  purgedCount: number;
  message: string;
}

export interface CdnStats {
  totalDomains: number;
  totalRequests: number;
  cacheHits: number;
  cacheMisses: number;
  cacheHitRate: number;
  /** Total bytes served */
  totalBandwidth: number;
  activeOrigins: number;
}

export interface DomainStats {
  domainId: number;
  requests: number;
  cacheHits: number;
  cacheMisses: number;
  cacheHitRate: number;
  bandwidth: number;
  cachedFiles: number;
  /** Cache size in bytes */
  cacheSize: number;
}