# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CDN Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mobile-auth/token-exchange:
    post:
      operationId: exchangeMobileAuthorizationCode
      tags: [auth]
      summary: Exchange mobile auth code for session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeBody"
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeMobileAuthorizationCodeResponse"
...
  /domains:
    get:
      operationId: listDomains
      tags: [domains]
      summary: List all domains
      responses:
        "200":
          description: List of domains
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Domain"
    post:
      operationId: createDomain
      tags: [domains]
      summary: Add a new domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDomainBody"
      responses:
        "201":
          description: Domain created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        "400":
          description: Validation error
        "403":
          description: Forbidden
  /domains/{domainId}:
    get:
      operationId: getDomain
      tags: [domains]
      summary: Get a domain by ID
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Domain details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DomainWithOrigins"
        "404":
          description: Not found
    patch:
      operationId: updateDomain
      tags: [domains]
      summary: Update a domain
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDomainBody"
      responses:
        "200":
          description: Updated domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Domain"
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteDomain
      tags: [domains]
      summary: Delete a domain
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /domains/{domainId}/cache/purge:
    post:
      operationId: purgeCache
      tags: [cache]
      summary: Purge cache for a domain
      parameters:
        - name: domainId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurgeCacheBody"
      responses:
        "200":
          description: Purge result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurgeCacheResponse"
        "403":
          description: Forbidden
...
components:
  schemas:
...
    ExchangeMobileAuthorizationCodeBody:
      type: object
      properties:
        code:
          type: string
        code_verifier:
          type: string
        redirect_uri:
          type: string
        state:
          type: string
        nonce:
          type: string
          nullable: true
      required:
        - code
        - code_verifier
        - redirect_uri
        - state
    ExchangeMobileAuthorizationCodeResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token
...
    Domain:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        sslEnabled:
          type: boolean
        cacheEnabled:
          type: boolean
        cacheTtl:
          type: integer
          description: Cache TTL in seconds
        status:
          type: string
          enum: [active, inactive, provisioning]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - sslEnabled
        - cacheEnabled
        - cacheTtl
        - status
        - createdAt
        - updatedAt
    DomainWithOrigins:
      allOf:
        - $ref: "#/components/schemas/Domain"
        - type: object
          properties:
            origins:
              type: array
              items:
                $ref: "#/components/schemas/Origin"
          required:
            - origins
    CreateDomainBody:
      type: object
      properties:
        name:
          type: string
        sslEnabled:
          type: boolean
        cacheEnabled:
          type: boolean
        cacheTtl:
          type: integer
      required:
        - name
    UpdateDomainBody:
      type: object
      properties:
        name:
          type: string
        sslEnabled:
          type: boolean
        cacheEnabled:
          type: boolean
        cacheTtl:
          type: integer
        status:
          type: string
          enum: [active, inactive, provisioning]
    Origin:
      type: object
      properties:
        id:
          type: integer
        domainId:
          type: integer
        address:
          type: string
          description: IP address or hostname of the origin server
        port:
          type: integer
        protocol:
          type: string
          enum: [http, https]
        weight:
          type: integer
          description: Load balancing weight (1-100)
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - domainId
        - address
        - port
        - protocol
        - weight
        - isActive
        - createdAt
...
    PurgeCacheBody:
      type: object
      properties:
        path:
          type: string
          description: Specific path to purge, or omit to purge all
    PurgeCacheResponse:
      type: object
      properties:
        purgedCount:
          type: integer
        message:
          type: string
      required:
        - purgedCount
        - message
...