# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CDN Admin API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get current authenticated user
 */
export const GetCurrentAuthUserResponse = zod.object({
  authenticated: zod.boolean(),
  user: zod
    .object({
      id: zod.string(),
      username: zod.string().nullish(),
      email: zod.string().nullish(),
      firstName: zod.string().nullish(),
      lastName: zod.string().nullish(),
      profileImage: zod.string().nullish(),
      profileImageUrl: zod.string().nullish(),
      role: zod.enum(["superadmin", "admin", "user"]),
    })
    .nullish(),
});

/**
 * @summary List all users (superadmin only)
 */
export const ListUsersResponseItem = zod.object({
  id: zod.string(),
  username: zod.string(),
  firstName: zod.string().optional(),
  lastName: zod.string().optional(),
  profileImage: zod.string().nullish(),
  role: zod.enum(["superadmin", "admin", "user"]),
  createdAt: zod.coerce.date(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Update a user role (superadmin only)
 */
export const UpdateUserParams = zod.object({
  userId: zod.coerce.string(),
});

export const UpdateUserBody = zod.object({
  role: zod.enum(["superadmin", "admin", "user"]),
});

export const UpdateUserResponse = zod.object({
  id: zod.string(),
  username: zod.string(),
  firstName: zod.string().optional(),
  lastName: zod.string().optional(),
  profileImage: zod.string().nullish(),
  role: zod.enum(["superadmin", "admin", "user"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Exchange mobile auth code for session token
 */
export const ExchangeMobileAuthorizationCodeBody = zod.object({
  code: zod.string(),
  code_verifier: zod.string(),
  redirect_uri: zod.string(),
  state: zod.string(),
  nonce: zod.string().nullish(),
});

export const ExchangeMobileAuthorizationCodeResponse = zod.object({
  token: zod.string(),
});

/**
 * @summary Logout mobile session
 */
export const LogoutMobileSessionResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List all domains
 */
export const ListDomainsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  sslEnabled: zod.boolean(),
  cacheEnabled: zod.boolean(),
  cacheTtl: zod.number().describe("Cache TTL in seconds"),
  status: zod.enum(["active", "inactive", "provisioning"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListDomainsResponse = zod.array(ListDomainsResponseItem);

/**
 * @summary Add a new domain
 */
export const CreateDomainBody = zod.object({
  name: zod.string(),
  sslEnabled: zod.boolean().optional(),
  cacheEnabled: zod.boolean().optional(),
  cacheTtl: zod.number().optional(),
});

/**
 * @summary Get a domain by ID
 */
export const GetDomainParams = zod.object({
  domainId: zod.coerce.number(),
});

export const GetDomainResponse = zod
  .object({
    id: zod.number(),
    name: zod.string(),
    sslEnabled: zod.boolean(),
    cacheEnabled: zod.boolean(),
    cacheTtl: zod.number().describe("Cache TTL in seconds"),
    status: zod.enum(["active", "inactive", "provisioning"]),
    createdAt: zod.coerce.date(),
    updatedAt: zod.coerce.date(),
  })
  .and(
    zod.object({
      origins: zod.array(
        zod.object({
          id: zod.number(),
          domainId: zod.number(),
          address: zod
            .string()
            .describe("IP address or hostname of the origin server"),
          port: zod.number(),
          protocol: zod.enum(["http", "https"]),
          weight: zod.number().describe("Load balancing weight (1-100)"),
          isActive: zod.boolean(),
          createdAt: zod.coerce.date(),
        }),
      ),
    }),
  );

/**
 * @summary Update a domain
 */
export const UpdateDomainParams = zod.object({
  domainId: zod.coerce.number(),
});

export const UpdateDomainBody = zod.object({
  name: zod.string().optional(),
  sslEnabled: zod.boolean().optional(),
  cacheEnabled: zod.boolean().optional(),
  cacheTtl: zod.number().optional(),
  status: zod.enum(["active", "inactive", "provisioning"]).optional(),
});

export const UpdateDomainResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  sslEnabled: zod.boolean(),
  cacheEnabled: zod.boolean(),
  cacheTtl: zod.number().describe("Cache TTL in seconds"),
  status: zod.enum(["active", "inactive", "provisioning"]),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a domain
 */
export const DeleteDomainParams = zod.object({
  domainId: zod.coerce.number(),
});

/**
 * @summary List origin servers for a domain
 */
export const ListOriginsParams = zod.object({
  domainId: zod.coerce.number(),
});

export const ListOriginsResponseItem = zod.object({
  id: zod.number(),
  domainId: zod.number(),
  address: zod.string().describe("IP address or hostname of the origin server"),
  port: zod.number(),
  protocol: zod.enum(["http", "https"]),
  weight: zod.number().describe("Load balancing weight (1-100)"),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListOriginsResponse = zod.array(ListOriginsResponseItem);

/**
 * @summary Add an origin server to a domain
 */
export const CreateOriginParams = zod.object({
  domainId: zod.coerce.number(),
});

export const CreateOriginBody = zod.object({
  address: zod.string(),
  port: zod.number().optional(),
  protocol: zod.enum(["http", "https"]).optional(),
  weight: zod.number().optional(),
});

/**
 * @summary Update an origin server
 */
export const UpdateOriginParams = zod.object({
  domainId: zod.coerce.number(),
  originId: zod.coerce.number(),
});

export const UpdateOriginBody = zod.object({
  address: zod.string().optional(),
  port: zod.number().optional(),
  protocol: zod.enum(["http", "https"]).optional(),
  weight: zod.number().optional(),
  isActive: zod.boolean().optional(),
});

export const UpdateOriginResponse = zod.object({
  id: zod.number(),
  domainId: zod.number(),
  address: zod.string().describe("IP address or hostname of the origin server"),
  port: zod.number(),
  protocol: zod.enum(["http", "https"]),
  weight: zod.number().describe("Load balancing weight (1-100)"),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete an origin server
 */
export const DeleteOriginParams = zod.object({
  domainId: zod.coerce.number(),
  originId: zod.coerce.number(),
});

/**
 * @summary Purge cache for a domain
 */
export const PurgeCacheParams = zod.object({
  domainId: zod.coerce.number(),
});

export const PurgeCacheBody = zod.object({
  path: zod
    .string()
    .optional()
    .describe("Specific path to purge, or omit to purge all"),
});

export const PurgeCacheResponse = zod.object({
  purgedCount: zod.number(),
  message: zod.string(),
});

/**
 * @summary Get overall CDN statistics
 */
export const GetStatsResponse = zod.object({
  totalDomains: zod.number(),
  totalRequests: zod.number(),
  cacheHits: zod.number(),
  cacheMisses: zod.number(),
  cacheHitRate: zod.number(),
  totalBandwidth: zod.number().describe("Total bytes served"),
  activeOrigins: zod.number(),
});

/**
 * @summary Get stats for a specific domain
 */
export const GetDomainStatsParams = zod.object({
  domainId: zod.coerce.number(),
});

export const GetDomainStatsResponse = zod.object({
  domainId: zod.number(),
  requests: zod.number(),
  cacheHits: zod.number(),
  cacheMisses: zod.number(),
  cacheHitRate: zod.number(),
  bandwidth: zod.number(),
  cachedFiles: zod.number(),
  cacheSize: zod.number().describe("Cache size in bytes"),
});