# lib/api-zod/src/generated/types/domain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CDN Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { DomainStatus } from "./domainStatus";

export interface Domain {
  id: number;
  name: string;
  sslEnabled: boolean;
  cacheEnabled: boolean;
  /** Cache TTL in seconds */
  cacheTtl: number;
  status: DomainStatus;
  createdAt: Date;
  updatedAt: Date;
}