# lib/api-zod/src/generated/types/domainWithOrigins.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CDN Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { Domain } from "./domain";
import type { Origin } from "./origin";

export type DomainWithOrigins = Domain & {
  origins: Origin[];
};