# lib/api-zod/src/generated/types/purgeCacheBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CDN Admin API
 * OpenAPI spec version: 0.1.0
 */

export interface PurgeCacheBody {
  /** Specific path to purge, or omit to purge all */
  path?: string;
}